✨ 수파베이스 타입 타일 총 정리
