+++ package.json
@@ -6,7 +6,9 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "server": "json-server --watch --port 5432 db.json",
+    "genTypes": "supabase gen types typescript --project-id dsggwbvtcrwuopwelpxy --schema public > src>types>supabase>supabase-type.ts "
   },
   "dependencies": {
     "@fullcalendar/core": "^6.1.15",
