+++ src/types/supabase/answer-type.ts
@@ -1,0 +1,3 @@
+import { Tables } from "./supabase-type";
+
+export type AnswerType = Tables<'answers'>;
