+++ src/types/supabase/booking-type.ts
@@ -1,0 +1,3 @@
+import { Tables } from "./supabase-type";
+
+export type BookingType = Tables<'bookings'>;
