+++ src/types/supabase/contact-type.ts
@@ -1,0 +1,3 @@
+import { Tables } from "./supabase-type";
+
+export type ContactType = Tables<'contacts'>;
