+++ src/types/supabase/hotel-type.ts
@@ -1,0 +1,3 @@
+import { Tables } from "./supabase-type";
+
+export type HotelType = Tables<'hotels'>;
