+++ src/types/supabase/review-type.ts
@@ -1,0 +1,3 @@
+import { Tables } from "./supabase-type";
+
+export type ReviewType = Tables<'reviews'>;
