+++ src/types/supabase/room-type.ts
@@ -1,0 +1,3 @@
+import { Tables } from "./supabase-type";
+
+export type RoomType = Tables<'rooms'>;
