+++ src/types/supabase/user-type.ts
@@ -1,0 +1,3 @@
+import { Tables } from "./supabase-type";
+
+export type UserType = Tables<'users'>;
